Switch from Yarn to pnpm
Normally, I don't think this would be worth the effort. However, the
"classic" version of Yarn depends on some deprecated Node.js packages.

There's now a "Yarn Modern" version, but it seems very funky and it was
more work and management than I wanted to upgrade it. Instead, I decided
to use the opportunity to swtich to pnpm to try it out instead.